requirements_txt.txt
# ETL GEODATA - Requirements
# Instalar com: pip install -r requirements.txt

# Manipulação de dados
pandas>=1.5.0

# Conexão Oracle
cx_Oracle>=8.3.0

# Conexão PostgreSQL
psycopg2-binary>=2.9.0

# ORM e SQL toolkit
SQLAlchemy>=1.4.0

# Utilitários adicionais
python-dateutil>=2.8.0

# Logging e configuração
python-decouple>=3.6

# Para desenvolvimento/debug (opcional)
ipython>=8.0.0

# Notas de instalação:
#
# ORACLE INSTANT CLIENT:
# - Baixar Oracle Instant Client do site da Oracle
# - Ubuntu: sudo apt-get install libaio1
# - Configurar LD_LIBRARY_PATH se necessário
# - Alternativamente: sudo apt-get install oracle-instantclient-basic
#
# POSTGRESQL:
# - Ubuntu: sudo apt-get install libpq-dev python3-dev
# - Se erro de compilação psycopg2: usar psycopg2-binary
#
# SISTEMA:
# - Python 3.8+ requerido
# - Ubuntu: sudo apt-get install python3-pip python3-venv
